+++ sale/src/main/java/kr/game/sale/controller/AdminController.java
@@ -23,41 +23,8 @@ import java.io.IOException;
 @RequiredArgsConstructor
 public class AdminController {
     private final AdminService adminService;
-    private IamportClient iamportClient;
-    @PostConstruct
-    public void init(){
-        this.iamportClient = new IamportClient(apiKey,secretKey);
-    }
-    @Value("${imp.api.key}")
-    private String apiKey;
-    @Value("${imp.api.secretkey}")
-    private String secretKey;
     @GetMapping
     public String adminForm(){
         return "admin/adminForm";
-    }
-
-    // 결제 요청
-    @PostMapping("/payment")
-    public @ResponseBody IamportResponse<Payment> validateIamport(@ModelAttribute PaymentDataForm paymentData){
-        String impUid = paymentData.getImpUid();
-        //결제된 정보 요청
-        IamportResponse<Payment> payment = iamportClient.paymentByImpUid(impUid);
-        log.info("결제 요청 응답. 결제 내역 = {}",impUid);
-        return payment;
-    }
-    // 결제 마무리
-    @PostMapping("/paymenttest")
-    public ResponseEntity<String> paymentComplete(@ModelAttribute PaymentDataForm paymentData,@ModelAttribute PaymentForm payment) throws IOException {
-        String orderNumber = paymentData.getMerchantUid();
-        try{
-            log.info("결제 성공 : 주문번호 {}",orderNumber);
-            return ResponseEntity.ok().build();
-        } catch(RuntimeException e){
-            log.info("주문 상품 환불 진행 : 주문 번호 {}",orderNumber);
-            String token = adminService.getToken(apiKey,secretKey);
-            adminService.refundRequest(token,orderNumber,e.getMessage());
-            return new ResponseEntity<>(e.getMessage(), HttpStatus.BAD_REQUEST);
-        }
     }
 }
