+++ sale/src/main/resources/templates/admin/adminForm.html
@@ -1,6 +1,6 @@
 <!DOCTYPE html>
 <html lang="en" xmlns:th="http://www.thymeleaf.org">
-<head th:replace="~{fragment/header :: common_header(~{::link})}">
+<head>
 </head>
 <body>
